requirements.txt
asgiref==3.7.2
Django==5.0.2
django-filter==24.2
djangorestframework==3.15.2
Markdown==3.6
mysqlclient==2.2.4
python-dotenv==1.0.1
sqlparse==0.4.4
typing_extensions==4.9.0
